Updated packages; Using configuration classes for EF Core.
