+++ src/Infrastructure/Data/Config/BasketConfiguration.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using Microsoft.eShopWeb.ApplicationCore.Entities.BasketAggregate;
+
+namespace Microsoft.eShopWeb.Infrastructure.Data.Config
+{
+    public class BasketConfiguration : IEntityTypeConfiguration<Basket>
+    {
+        public void Configure(EntityTypeBuilder<Basket> builder)
+        {
+            var navigation = builder.Metadata.FindNavigation(nameof(Basket.Items));
+            navigation.SetPropertyAccessMode(PropertyAccessMode.Field);
+
+            builder.Property(b => b.BuyerId)
+                .IsRequired()
+                .HasMaxLength(20);
+        }
+    }
+}
