+++ src/Infrastructure/Data/Config/BasketItemConfiguration.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using Microsoft.eShopWeb.ApplicationCore.Entities.BasketAggregate;
+
+namespace Microsoft.eShopWeb.Infrastructure.Data.Config
+{
+    public class BasketItemConfiguration : IEntityTypeConfiguration<BasketItem>
+    {
+        public void Configure(EntityTypeBuilder<BasketItem> builder)
+        {
+            builder.Property(bi => bi.UnitPrice)
+                .IsRequired(true)
+                .HasColumnType("decimal(18,2)");
+        }
+    }
+}
