+++ src/Infrastructure/Data/Config/CatalogBrandConfiguration.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using Microsoft.eShopWeb.ApplicationCore.Entities;
+
+namespace Microsoft.eShopWeb.Infrastructure.Data.Config
+{
+    public class CatalogBrandConfiguration : IEntityTypeConfiguration<CatalogBrand>
+    {
+        public void Configure(EntityTypeBuilder<CatalogBrand> builder)
+        {
+            builder.HasKey(ci => ci.Id);
+
+            builder.Property(ci => ci.Id)
+               .ForSqlServerUseSequenceHiLo("catalog_brand_hilo")
+               .IsRequired();
+
+            builder.Property(cb => cb.Brand)
+                .IsRequired()
+                .HasMaxLength(100);
+        }
+    }
+}
