+++ src/Infrastructure/Data/Config/CatalogTypeConfiguration.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using Microsoft.eShopWeb.ApplicationCore.Entities;
+
+namespace Microsoft.eShopWeb.Infrastructure.Data.Config
+{
+    public class CatalogTypeConfiguration : IEntityTypeConfiguration<CatalogType>
+    {
+        public void Configure(EntityTypeBuilder<CatalogType> builder)
+        {
+            builder.HasKey(ci => ci.Id);
+
+            builder.Property(ci => ci.Id)
+               .ForSqlServerUseSequenceHiLo("catalog_type_hilo")
+               .IsRequired();
+
+            builder.Property(cb => cb.Type)
+                .IsRequired()
+                .HasMaxLength(100);
+        }
+    }
+}
