+++ src/Infrastructure/Data/Config/OrderConfiguration.cs
@@ -1,0 +1,18 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using Microsoft.eShopWeb.ApplicationCore.Entities.OrderAggregate;
+
+namespace Microsoft.eShopWeb.Infrastructure.Data.Config
+{
+    public class OrderConfiguration : IEntityTypeConfiguration<Order>
+    {
+        public void Configure(EntityTypeBuilder<Order> builder)
+        {
+            var navigation = builder.Metadata.FindNavigation(nameof(Order.OrderItems));
+
+            navigation.SetPropertyAccessMode(PropertyAccessMode.Field);
+
+            builder.OwnsOne(o => o.ShipToAddress);
+        }
+    }
+}
